+++ testmodul/isilive_test_fixtures.lua
@@ -101,6 +101,9 @@ local function BuildEventHandlersBaseOptions(entryRef, counters)
     updateLeaderButtons = function() end,
     updateStatusLine = function() end,
     sendIsiLiveHello = function(_force, _source) end,
+    sendLibKeystonePartyData = function(_force)
+      return false
+    end,
     notifyPostChallengeSync = function() end,
     sendOwnKeySnapshot = function(_force) end,
     sendOwnBackgroundSnapshot = function(_source) end,